LibWeb: Restore realm retrieval from 'this' in activate_event_handler()
Removing the FIXME'd code in fcefad4 was a bit too eager, and relying on
the main thread VM's current realm only works when JS is being executed.

Restore a simplified version of the old code to determine the realm this
time instead of the global object, following the assumptions already
made in get_current_value_of_event_handler() regarding what kind of
event target 'this' can be.